+++ modules/email-notifications/test/email-notification-service-test.js
@@ -226,6 +226,7 @@ describe('email-notification-service', function() {
       return emailNotificationService.sendInvitation(invitingUser, invite, room).then(function() {
         assert.strictEqual(emailPayload.subject, 'bob invited you to join the a/b chat on Gitter');
         assert.strictEqual(emailPayload.from, 'bob <support@gitter.im>');
+        assert.strictEqual(emailPayload.replyTo, 'no-reply@gitter.im');
         assert.strictEqual(emailPayload.to, 'x@troupetest.local');
         assert.strictEqual(emailPayload.data.roomUrl, BASE_EMAIL_PATH + '/a/b');
         assert.strictEqual(