+++ modules/gitlab/lib/index.js
@@ -1,6 +1,8 @@
 'use strict';
 
 module.exports = {
+  GitLabUriValidator: require('./gitlab-uri-validator'),
+  GitLabUserService: require('./user-service'),
   GitLabIssuableService: require('./issuable-service'),
   GitLabIssuableStateService: require('./issuable-state-service'),
   GitLabGroupService: require('./group-service')