+++ modules/permissions/test/ensure-access-and-fetch-descriptor-test.js
@@ -2,50 +2,49 @@
 
 var assert = require('assert');
 var fixtureLoader = require('gitter-web-test-utils/lib/test-fixtures');
-var StatusError = require('statuserror');
 var ensureAccessAndFetchDescriptor = require('gitter-web-permissions/lib/ensure-access-and-fetch-descriptor');
 
 describe('ensure-access-and-fetch-descriptor #slow', function() {
   this.timeout(10000);
 
   fixtureLoader.disableMongoTableScans();
-  fixtureLoader.ensureIntegrationEnvironment('#integrationUser1');
+  fixtureLoader.ensureIntegrationEnvironment('#integrationUser1', '#integrationGitlabUser1');
 
   var fixture = fixtureLoader.setup({
-    user1: '#integrationUser1'
+    user1: '#integrationUser1',
+    userGitlab1: '#integrationGitlabUser1'
   });
 
-  it('should return a descriptor for type null', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
+  it('should return a descriptor for type null', async () => {
+    const sd = await ensureAccessAndFetchDescriptor(fixture.user1, {
       security: 'PUBLIC'
-    }).then(function(sd) {
-      assert.deepEqual(sd, {
-        type: null,
-        admins: 'MANUAL',
-        public: true,
-        members: 'PUBLIC',
-        extraMembers: [],
-        extraAdmins: [fixture.user1._id]
-      });
+    });
+
+    assert.deepEqual(sd, {
+      type: null,
+      admins: 'MANUAL',
+      public: true,
+      members: 'PUBLIC',
+      extraMembers: [],
+      extraAdmins: [fixture.user1._id]
     });
   });
 
-  it('should throw an error if you give it an unknown type', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, { type: 'foo' })
-      .then(function() {
-        assert.ok(false, 'Expected error');
-      })
-      .catch(StatusError, function(err) {
-        assert.strictEqual(err.status, 400);
-      });
+  it('should throw an error if you give it an unknown type', async () => {
+    await assert.rejects(ensureAccessAndFetchDescriptor(fixture.user1, { type: 'foo' }), err => {
+      assert.strictEqual(err.status, 400);
+      return true;
+    });
   });
 
-  it('should return a descriptor for a github org if the user has access', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_ORG',
-      linkPath: fixtureLoader.GITTER_INTEGRATION_ORG,
-      security: 'PUBLIC'
-    }).then(function(sd) {
+  describe('GitHub', () => {
+    it('should return a descriptor for a github org if the user has access', async () => {
+      const sd = await ensureAccessAndFetchDescriptor(fixture.user1, {
+        type: 'GH_ORG',
+        linkPath: fixtureLoader.GITTER_INTEGRATION_ORG,
+        security: 'PUBLIC'
+      });
+
       assert.deepEqual(sd, {
         type: 'GH_ORG',
         members: 'PUBLIC',
@@ -55,14 +54,14 @@ describe('ensure-access-and-fetch-descriptor #slow', function() {
         externalId: fixtureLoader.GITTER_INTEGRATION_ORG_ID
       });
     });
-  });
 
-  it('should return a descriptor for a github user if the user has access', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_USER',
-      linkPath: fixtureLoader.GITTER_INTEGRATION_USERNAME,
-      security: 'PUBLIC'
-    }).then(function(sd) {
+    it('should return a descriptor for a github user if the user has access', async () => {
+      const sd = await ensureAccessAndFetchDescriptor(fixture.user1, {
+        type: 'GH_USER',
+        linkPath: fixtureLoader.GITTER_INTEGRATION_USERNAME,
+        security: 'PUBLIC'
+      });
+
       assert.deepEqual(sd, {
         type: 'GH_USER',
         members: 'PUBLIC',
@@ -73,16 +72,16 @@ describe('ensure-access-and-fetch-descriptor #slow', function() {
         extraAdmins: [] // no user id because the username matches
       });
     });
-  });
 
-  it('should return a descriptor for a github repo if the user has access', function() {
-    var linkPath = fixtureLoader.GITTER_INTEGRATION_REPO_FULL;
+    it('should return a descriptor for a github repo if the user has access', async () => {
+      const linkPath = fixtureLoader.GITTER_INTEGRATION_REPO_FULL;
 
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_REPO',
-      linkPath: linkPath,
-      security: 'PUBLIC'
-    }).then(function(sd) {
+      const sd = await ensureAccessAndFetchDescriptor(fixture.user1, {
+        type: 'GH_REPO',
+        linkPath: linkPath,
+        security: 'PUBLIC'
+      });
+
       assert.deepEqual(sd, {
         type: 'GH_REPO',
         members: 'PUBLIC',
@@ -92,18 +91,18 @@ describe('ensure-access-and-fetch-descriptor #slow', function() {
         externalId: fixtureLoader.GITTER_INTEGRATION_REPO_ID
       });
     });
-  });
 
-  it('should return a descriptor for an unknown github owner if the user has access', function() {
-    // suppose you don't know if linkPath is an org or user and you're just
-    // upserting a group for it
-    var linkPath = fixtureLoader.GITTER_INTEGRATION_ORG;
+    it('should return a descriptor for an unknown github owner if the user has access', async () => {
+      // suppose you don't know if linkPath is an org or user and you're just
+      // upserting a group for it
+      const linkPath = fixtureLoader.GITTER_INTEGRATION_ORG;
 
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_GUESS',
-      linkPath: linkPath,
-      security: 'PUBLIC'
-    }).then(function(sd) {
+      const sd = await ensureAccessAndFetchDescriptor(fixture.user1, {
+        type: 'GH_GUESS',
+        linkPath: linkPath,
+        security: 'PUBLIC'
+      });
+
       assert.deepEqual(sd, {
         type: 'GH_ORG',
         members: 'PUBLIC',
@@ -113,47 +112,108 @@ describe('ensure-access-and-fetch-descriptor #slow', function() {
         externalId: fixtureLoader.GITTER_INTEGRATION_ORG_ID
       });
     });
+
+    it('should throw an error if the returned github type does not match as expected', async () => {
+      await assert.rejects(
+        ensureAccessAndFetchDescriptor(fixture.user1, {
+          type: 'GH_ORG',
+          linkPath: fixtureLoader.GITTER_INTEGRATION_REPO_FULL,
+          security: 'PUBLIC'
+        }),
+        err => {
+          assert.strictEqual(err.status, 400);
+          return true;
+        }
+      );
+    });
+
+    it('should throw an error if the user does not have access', async () => {
+      assert.rejects(
+        ensureAccessAndFetchDescriptor(fixture.user1, {
+          type: 'GH_ORG',
+          linkPath: 'gitterHQ',
+          security: 'PUBLIC'
+        }),
+        err => {
+          assert.strictEqual(err.status, 403);
+          return true;
+        }
+      );
+    });
+
+    it('should throw an error if the github object does not exist', async () => {
+      assert.rejects(
+        ensureAccessAndFetchDescriptor(fixture.user1, {
+          type: 'GH_ORG',
+          linkPath: 'foo-foo-does-not-exist',
+          security: 'PUBLIC'
+        }),
+        err => {
+          assert.strictEqual(err.status, 404);
+          return true;
+        }
+      );
+    });
   });
 
-  it('should throw an error if the returned github type does not match as expected', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_ORG',
-      linkPath: fixtureLoader.GITTER_INTEGRATION_REPO_FULL,
-      security: 'PUBLIC'
-    })
-      .then(function() {
-        assert.ok(false, 'Expected error');
-      })
-      .catch(StatusError, function(err) {
-        assert.strictEqual(err.status, 400);
+  describe('GitLab', () => {
+    it('should return a descriptor for a GitLab group if the user has access', async () => {
+      const sd = await ensureAccessAndFetchDescriptor(fixture.userGitlab1, {
+        type: 'GL_GROUP',
+        linkPath: fixtureLoader.GITLAB_GROUP1_URI,
+        security: 'PUBLIC'
       });
-  });
 
-  it('should throw an error if the user does not have access', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_ORG',
-      linkPath: 'gitterHQ',
-      security: 'PUBLIC'
-    })
-      .then(function() {
-        assert.ok(false, 'Expected error');
-      })
-      .catch(StatusError, function(err) {
-        assert.strictEqual(err.status, 403);
+      assert.deepEqual(sd, {
+        type: 'GL_GROUP',
+        members: 'PUBLIC',
+        admins: 'GL_GROUP_MAINTAINER',
+        public: true,
+        linkPath: fixtureLoader.GITLAB_GROUP1_URI,
+        externalId: fixtureLoader.GITLAB_GROUP1_ID
       });
-  });
+    });
 
-  it('should throw an error if the github object does not exist', function() {
-    return ensureAccessAndFetchDescriptor(fixture.user1, {
-      type: 'GH_ORG',
-      linkPath: 'foo-foo-does-not-exist',
-      security: 'PUBLIC'
-    })
-      .then(function() {
-        assert.ok(false, 'Expected error');
-      })
-      .catch(StatusError, function(err) {
-        assert.strictEqual(err.status, 404);
-      });
+    it('should throw an error if the returned GitLab type does not match as expected', async () => {
+      assert.rejects(
+        ensureAccessAndFetchDescriptor(fixture.userGitlab1, {
+          type: 'GL_GROUP',
+          linkPath: fixtureLoader.GITLAB_USER_USERNAME,
+          security: 'PUBLIC'
+        }),
+        err => {
+          assert.strictEqual(err.status, 400);
+          return true;
+        }
+      );
+    });
+
+    it('should throw an error if the user does not have access', async () => {
+      assert.rejects(
+        ensureAccessAndFetchDescriptor(fixture.userGitlab1, {
+          type: 'GL_GROUP',
+          linkPath: 'gitlab-org/gitter',
+          security: 'PUBLIC'
+        }),
+        err => {
+          assert.strictEqual(err.status, 403);
+          return true;
+        }
+      );
+    });
+
+    it('should throw an error if the GitLab object does not exist', async () => {
+      assert.rejects(
+        ensureAccessAndFetchDescriptor(fixture.userGitlab1, {
+          type: 'GL_GROUP',
+          linkPath: 'foo-foo-does-not-exist',
+          security: 'PUBLIC'
+        }),
+        err => {
+          assert.strictEqual(err.status, 404);
+          return true;
+        }
+      );
+    });
   });
 });