+++ server/services/restful.js
@@ -176,26 +176,8 @@ function serializeOrgsForUserId(userId, options) {
   });
 }
 
-function serializeUnusedOrgsForUser(user) {
-  return orgService.getUnusedOrgsForUser(user).then(function(orgs) {
-    var strategy = new restSerializer.GithubOrgStrategy({
-      currentUserId: user && user._id
-    });
-    return restSerializer.serialize(orgs, strategy);
-  });
-}
-
 function serializeReposForUser(user) {
   return repoService.getReposForUser(user).then(function(repos) {
-    var strategy = new restSerializer.GithubRepoStrategy({
-      currentUserId: user && user._id
-    });
-    return restSerializer.serialize(repos, strategy);
-  });
-}
-
-function serializeUnusedReposForUser(user) {
-  return repoService.getUnusedReposForUser(user).then(function(repos) {
     var strategy = new restSerializer.GithubRepoStrategy({
       currentUserId: user && user._id
     });
@@ -280,9 +262,7 @@ module.exports = {
   serializeEventsForTroupe: serializeEventsForTroupe,
   serializeOrgsForUser: serializeOrgsForUser,
   serializeOrgsForUserId: serializeOrgsForUserId,
-  serializeUnusedOrgsForUser: serializeUnusedOrgsForUser,
   serializeReposForUser: serializeReposForUser,
-  serializeUnusedReposForUser: serializeUnusedReposForUser,
   serializeAdminReposForUser: serializeAdminReposForUser,
   serializeProfileForUsername: serializeProfileForUsername,
   serializeGroupsForUserId: Promise.method(serializeGroupsForUserId),