+++ test/request-api-tests/user-repos-api-tests.js
@@ -4,7 +4,6 @@ process.env.DISABLE_API_LISTEN = '1';
 
 var assert = require('assert');
 var fixtureLoader = require('gitter-web-test-utils/lib/test-fixtures');
-var groupService = require('gitter-web-groups/lib/group-service');
 
 describe('user-repos #slow', function() {
   var app, request;
@@ -50,48 +49,6 @@ describe('user-repos #slow', function() {
         assert(
           repos.some(function(repo) {
             return repo.name === fixtureLoader.GITTER_INTEGRATION_REPO_FULL;
-          })
-        );
-      });
-  });
-
-  it('GET /v1/user/:userId/repos?type=unused', function() {
-    return request(app)
-      .get('/v1/user/' + fixture.user1.id + '/repos?type=unused')
-      .set('x-access-token', fixture.user1.accessToken)
-      .expect(200)
-      .then(function(result) {
-        var repos = result.body;
-
-        assert(
-          repos.some(function(repo) {
-            return repo.name === fixtureLoader.GITTER_INTEGRATION_REPO_FULL;
-          })
-        );
-
-        // now try and add one and see if it is still in there
-        // (should we do this via the API too? Going with groupService directly
-        //  as it is faster to execute and which user took the linkPath is
-        //  irrelevant)
-        return groupService.createGroup(fixture.user1, {
-          type: 'GH_REPO',
-          name: fixtureLoader.GITTER_INTEGRATION_COMMUNITY,
-          uri: fixtureLoader.GITTER_INTEGRATION_COMMUNITY,
-          linkPath: fixtureLoader.GITTER_INTEGRATION_REPO_FULL
-        });
-      })
-      .then(function() {
-        return request(app)
-          .get('/v1/user/' + fixture.user1.id + '/repos?type=unused')
-          .set('x-access-token', fixture.user1.accessToken)
-          .expect(200);
-      })
-      .then(function(result) {
-        var repos = result.body;
-
-        assert(
-          repos.every(function(repo) {
-            return repo.name !== fixtureLoader.GITTER_INTEGRATION_REPO_FULL;
           })
         );
       });